fixes: restructure files, actualize README.MD
