+++ src/main/kotlin/trees/UnbalancedBSTree.kt
@@ -1,7 +1,7 @@
-package trees.unbalancedBST
+package trees
 
 import nodes.UnbalancedBSTNode
-import trees.AbstractTree
+import trees.abstractTree.AbstractTree
 
 class UnbalancedBSTree<K : Comparable<K>, V> : AbstractTree<K, V, UnbalancedBSTNode<K, V>>() {
